# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic API"
  description: "This is an API to access Service Tickets and Inventory for a Mechanic Shop"
  version: "1.0.0"
host: "my-mechanic-helper.onrender.com" #changed from local 
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions: #Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey #double check this later
    name: Authorization
    in: header
paths: 
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: "Login endpoint and generates token"
      description: "Validates Mechanic credentials (email and password), and creates a mechanic specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login Successful"
          schema:
            $ref: "#/definitions/LoginSuccessResponse"
          examples:
            application/json:
              message: "Welcome Mechanic"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOj..."
              mechanic: 
                id: 0
                firstname: "Test"
                lastname: "Mech"
                email: "test@email.com"
                password: "<encrypted password>"
                salary: "123000"
                address: "123 Fun St."

        401:
          description: "Login Failed"
          schema:
            $ref: "#/definitions/LoginFailResponse"
          examples:
            application/json:
              message: "Failed to find Mechanic credentials"
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: "Register Mechanic"
      description: "Send in Mechanic info, to create a Mechanic profile that will be stored in the database. Address is an optional field."
      parameters:
        - in: "body"
          name: "body"
          description: "New Mechanic information"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"

      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 0
              firstname: "Test"
              lastname: "Mech"
              email: "test@email.com"
              password: "<encrypted password>"
              salary: "123000"
              address: "123 Fun St."

    
    get:
      tags:
        - Mechanics
      summary: "Returns all mechanics"
      description: "Endpoint to receive an array of mechanic objects."
      responses:
        200:
          description: "Successfully Retrieved Mechanics"
          schema:
            $ref: "#/definitions/MechanicsResponse"

    put:
      tags:
        - Mechanics
      summary: "Updates a Mechanic"
      description: "Endpoint to update a Mechanic profile. Requires a token. Mechanic will be identified by token."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic Updates"
          required: true
          schema:
            $ref: "#/definitions/MechanicInput"
  
      responses:
        200:
          description: "Successfully updated Mechanic profile."
          schema:
            $ref: "#/definitions/MechanicResponse"
          examples:
            application/json:
              id: 0
              firstname: "Test"
              lastname: "Mech"
              email: "test@email.com"
              password: "<encrypted password>"
              salary: "123000"
              address: "123 Fun St."
    
    delete:
      tags:
        - Mechanics
      summary: "Deletes a Mechanic"
      description: "Deletes a Mechanic object. Requires a token. Token will be used to identify Mechanic to delete."
      security:
        - bearerAuth: []

      responses:
        200:
          description: "Successfully deleted Mechanic profile."
          schema:
            type: object
            properties:
              message:
                type: string
  /mechanics/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: "Returns a mechanic"
      description: "Endpoint to receive a Mechanic object by using the Mechanic's id."
      parameters:
        - in: path
          name: mechanic_id
          required: true
          type: integer
          description: Numeric ID of the Mechanic to retrieve
      responses:
        200:
          description: "Successfully Retrieved Mechanic"
          schema:
            $ref: "#/definitions/MechanicResponse"
...
definitions: 
...
  LoginCredentials: #Input data required for login route
    type: object
    properties:
      email:
        type: string
      password:
        type: string

    required:
      - email
      - password
  LoginSuccessResponse: #Response from login endpoint
    type: object
    properties:
      message:
        type: string
      token:
        type: string
      mechanic:
       $ref: '#/definitions/MechanicResponse'
  LoginFailResponse:
    type: object
    properties:
      message:
        type: string

# -------------------

#-------------------Mechanic Input---------------
  MechanicInput: #Sending user Information
    type: object
    properties:
      firstname:
        type: string
        example: "Test"
      lastname:
        type: string
        example: "Mech"
      email:
        type: string
        example: "test@email.com"
      password:
        type: string
        example: "123"
      salary:
        type: string
        example: "123000" 
      address:
        type: string
        example: "123 Fun St."
      
    required:
      - firstname
      - lastname
      - email
      - password
      - salary
  MechanicResponse: 
    type: object
    properties:
      id:
        type: integer
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      password:
        type: string
      salary:
        type: string
      address:
        type: string
  MechanicsResponse: 
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        password:
          type: string
        salary:
          type: string
        address:
          type: string
#--------------------

#-------------------Mechanic Tickets---------------
...